pyproject.toml
[project]
name = "myfirstagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "ddgs>=9.10.0",
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.1",
]
